Crosshair rendering decoupled from WorldRenderer class
